Refactor identifyTLF out of KBFSNameInfoSource
